requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
groq>=0.11.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.36
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.6.0
email-validator>=2.0.0
